Coding style fixes
Codiung style fixes to help merge in future
TEST-180 Functional tests: pick media server logs
TEST-180 Functional tests: pick media server logs - bugfix
TEST-180 Functional tests: use official ubuntu box for vagrant images; fix media stream tests
TEST-180: fix resource synchronization test
TEST-180: time synchronization test, fixes
VMS-5860: More cameras from DW report
VMS-5860: Server incorrectly configures DW onvif camera
VMS-5870 PCI Scan Report
VMS-5882: Can't configure local SMTP Server as Email Server if there's no Internet connection
VMS-5893: There's no advanced settings for DW
func_tests: temporary disable nat tests
vms-4372. Cmake adjustments.
vms-4372. In progress.
vms-4372. Merge dev_3.0.0 --> vms_4372.
vms-4372. Merge dev_3.0.0 --> x_vms_4372.
vms-4372. Merge dev_3.0.0 <-- vms_4372.
vms-4372. Merge vms-4372 <-- dev_3.0.0.
vms-4372. Merge vms-4372 in progress.
vms-4372. Merge vms_4372 <-- dev_3.0.0.
vms-4372. Review fixes.
vms-4372. Test storage plugin in progress.
vms-4372. Test storage plugin.
vms-4372. UT in progress.
vms-4372. Windows build fixes.
vms-4372. Windows fixes.
vms-4372. in progress.
vms-5846. Server 'blinking' after merge resolved.
vms-5875. Mounted NAS after server restart issue resolved.
vms-5876. Incorrect spaceLimit for hand-mounted storage issue resolved.
vms. Wrong assert removed.

<==== Mercurial data ====>
Branch: dev_3.0.0_gui

+++ func_tests/test_utils/media_stream.py
@@ -172,7 +172,7 @@ class DirectHlsMediaStream(object):
         assert isinstance(pos, datetime), repr(pos)
         assert isinstance(duration, timedelta), repr(duration)
         temp_file_path = temp_file_name + '.%s' % self.format
-        pos_ms = datetime_utc_to_timestamp(pos) * 1000
+        pos_ms = int(datetime_utc_to_timestamp(pos) * 1000)
         duration_sec = int(duration.total_seconds() + 1)  # round to next value
         params = dict(pos=pos_ms, duration=duration_sec)
         metadata = load_stream_metadata_from_http(
